1. Added the base code for the random enpoint plugin.
This plugin helps send users to a random page on the website. Sending users to a random page or random set of pages will help shape traffic.

This plugin should be as system agnogstic as possible. While this inital plugin is targeted at wordpress this plugin might be used for another system in the future.